Extracted Lookup from the Numba code (surprisingly easy).
